+++ src/secrets.js
@@ -34,9 +34,17 @@ async function getSecrets(secretRequests, client) {
             body = responseCache.get(requestPath);
             cachedResponse = true;
         } else {
-            const result = await client.get(requestPath);
-            body = result.body;
-            responseCache.set(requestPath, body);
+            try {
+                const result = await client.get(requestPath);
+                body = result.body;
+                responseCache.set(requestPath, body);
+            } catch (error) {
+                const {response} = error;
+                if (response.statusCode === 404) {
+                    throw Error(`Unable to retrieve result for "${path}" because it was not found: ${response.body.trim()}`)
+                }
+                throw error
+            }
         }
 
         if (selector == wildcard) {                     
@@ -96,7 +104,8 @@ async function getSecrets(secretRequests, client) {
                 selector = "data." + selector
             }
 
-            const value = selectData(body, selector);
+
+            const value = await selectData(body, selector);
             results.push({
                 request: secretRequest,
                 value,
@@ -113,12 +122,12 @@ async function getSecrets(secretRequests, client) {
  * @param {object} data 
  * @param {string} selector 
  */
-function selectData(data, selector) {
+async function selectData(data, selector) {
     const ata = jsonata(selector);
-    let result = JSON.stringify(ata.evaluate(data));
+    let result = JSON.stringify(await ata.evaluate(data));
     // Compat for custom engines
     if (!result && ((ata.ast().type === "path" && ata.ast()['steps'].length === 1) || ata.ast().type === "string") && selector !== 'data' && 'data' in data) {
-        result = JSON.stringify(jsonata(`data.${selector}`).evaluate(data));
+        result = JSON.stringify(await jsonata(`data.${selector}`).evaluate(data));
     } else if (!result) {
         throw Error(`Unable to retrieve result for ${selector}. No match data was found. Double check your Key or Selector.`);
     }